atoms/getting_started/pyproject.toml
[project]
name = "atoms-getting-started"
version = "0.1.0"
description = "Getting started with the Atoms SDK - outbound calls and analytics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "smallestai>=4.2.0",
    "python-dotenv",
    "requests",
    "loguru",
]
